+++ CHANGELOG.md
@@ -8,17 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed
+- Added missing dist files
 
 ### Added
 - Server now has a delay configuration - this waits for N ms before returning the results. This is to simulate boat normal and abnormal response times.
 - Password generator MD5 hash function
 - Password generator has Salt function
 - Authentication generator has a JWT function
+- Added an alpha numeric generator
+- Added a 'from sets' function in the maths number generator
 
 ### Changed
-
-### Fixed
-- Added missing dist files
+- Removed docs for now as they weren't working correctly
+- Removed performance scripts as they're just cluttering up the repo - will probably add them back in at a later date, or make a clean repo for it.
 
 ## 0.0.4 - 2023-01-06
 ### Added